Renamed some compoenents
